common: add missing group and block metrics to stream
