+++ common/src/data_stream/mod.rs
@@ -1,5 +1,6 @@
 mod filter;
 mod fragment_access;
+mod metrics;
 mod segment_access;
 mod segment_stream;
 mod stream;
@@ -7,6 +8,7 @@ mod stream_group;
 
 pub use self::filter::{BlockFilterFactory, FilterMatch};
 pub use self::fragment_access::FragmentAccess;
+pub use self::metrics::DataStreamMetrics;
 pub use self::segment_access::{SegmentAccess, SegmentAccessFetch};
 pub use self::segment_stream::SegmentStream;
-pub use self::stream::{DataStream, DataStreamError, DataStreamMetrics};
+pub use self::stream::{DataStream, DataStreamError};
