+++ common/src/data_stream/stream.rs
@@ -4,7 +4,7 @@ use apibara_dna_protocol::dna::stream::{
     stream_data_response::Message, Data, DataFinality, DataProduction, Finalize, Invalidate,
     StreamDataResponse,
 };
-use apibara_observability::{Histogram, KeyValue, RecordRequest, RequestMetrics, UpDownCounter};
+use apibara_observability::{KeyValue, RecordRequest};
 use bytes::{BufMut, Bytes, BytesMut};
 use error_stack::{Result, ResultExt};
 use futures::FutureExt;
@@ -24,16 +24,10 @@ use crate::{
     Cursor,
 };
 
+use super::DataStreamMetrics;
+
 #[derive(Debug)]
 pub struct DataStreamError;
-
-#[derive(Debug, Clone)]
-pub struct DataStreamMetrics {
-    pub active: UpDownCounter<i64>,
-    pub block_size: Histogram<u64>,
-    pub fragment_size: Histogram<u64>,
-    pub segment: RequestMetrics,
-}
 
 pub struct DataStream {
     block_filter: Vec<BlockFilter>,
@@ -202,6 +196,7 @@ impl DataStream {
             self.fragment_id_to_name.clone(),
             self.store.clone(),
             self.chain_view.clone(),
+            self.metrics.clone(),
         );
 
         let (segment_tx, segment_rx) = mpsc::channel(SEGMENT_CHANNEL_SIZE);
@@ -301,6 +296,7 @@ impl DataStream {
         let block_entry: BlockAccess = self
             .store
             .get_block(&cursor)
+            .record_request(self.metrics.block.clone())
             .await
             .map_err(FileCacheError::Foyer)
             .change_context(DataStreamError)
@@ -674,62 +670,6 @@ impl error_stack::Context for DataStreamError {}
 impl std::fmt::Display for DataStreamError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "data stream error")
-    }
-}
-
-impl Default for DataStreamMetrics {
-    fn default() -> Self {
-        let meter = apibara_observability::meter("dna_data_stream");
-
-        Self {
-            active: meter
-                .i64_up_down_counter("dna.data_stream.active")
-                .with_description("number of active data streams")
-                .with_unit("{connection}")
-                .build(),
-            block_size: meter
-                .u64_histogram("dna.data_stream.block_size")
-                .with_description("size (in bytes) of blocks sent to the client")
-                .with_unit("By")
-                .with_boundaries(vec![
-                    1_000.0,
-                    10_000.0,
-                    100_000.0,
-                    1_000_000.0,
-                    5_000_000.0,
-                    10_000_000.0,
-                    25_000_000.0,
-                    50_000_000.0,
-                    100_000_000.0,
-                    1_000_000_000.0,
-                ])
-                .build(),
-            fragment_size: meter
-                .u64_histogram("dna.data_stream.fragment_size")
-                .with_description("size (in bytes) of fragments sent to the client")
-                .with_unit("By")
-                .with_boundaries(vec![
-                    1_000.0,
-                    10_000.0,
-                    100_000.0,
-                    1_000_000.0,
-                    5_000_000.0,
-                    10_000_000.0,
-                    25_000_000.0,
-                    50_000_000.0,
-                    100_000_000.0,
-                    1_000_000_000.0,
-                ])
-                .build(),
-            segment: RequestMetrics::new_with_boundaries(
-                "dna_data_stream",
-                "dna.data_stream.segment",
-                vec![
-                    0.0001, 0.00025, 0.0005, 0.001, 0.0025, 0.005, 0.0075, 0.01, 0.025, 0.05,
-                    0.075, 0.1, 0.25, 0.5, 0.75, 1.0, 2.5, 5.0, 10.0, 20.0, 30.0, 60.0, 120.0,
-                ],
-            ),
-        }
     }
 }
